Restructured folder & implemented color change in dots
