+++ src/client/components/LeftArrowComponent/LeftArrowComponent.js
