+++ src/client/components/LeftArrowComponent/LeftArrowComponent.stories.js
