+++ src/client/components/LeftArrowComponent/LeftArrowComponent.styles.css
