+++ src/client/components/RightArrowComponent/RightArrowComponent.js
