+++ src/client/components/RightArrowComponent/RightArrowComponent.stories.js
