+++ src/client/components/RightArrowComponent/RightArrowComponent.styles.css
