+++ src/client/components/SliderComponent/SliderComponent.stories.js
