+++ src/client/components/SliderComponent/SliderComponent.styles.css
