+++ src/client/components/SliderDotsComponent/SliderDotsComponent.styles.css
