+++ src/client/components/SlideshowImageComponent/SlideshowImageComponent.styles.css
