+++ src/client/components/SlideshowImageComponent/slideshowImageComponent.stories.js
@@ -5,5 +5,7 @@ import { withKnobs, number } from '@storybook/addon-knobs';
 export default { title: 'Avatars Component', decorators: [withKnobs] };
 
 export const Avatars = () => {
-  return <SlideshowImageComponent currentImageIndex={number('imageIndex', 0)} />;
+  return (
+    <SlideshowImageComponent currentImageIndex={number('imageIndex', 0)} />
+  );
 };
